+++ Source/Core/GZBuilder/Data/GlowingFlatData.cs
@@ -5,7 +5,7 @@ namespace CodeImp.DoomBuilder.GZBuilder.Data
 	public class GlowingFlatData
 	{
 		public PixelColor Color;
-		public float Height;
+		public double Height;
 		public int Brightness = 255;
 		public bool Fullbright;
 		public bool CalculateTextureColor;
